% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rgmdn.R
\name{gmdn}
\alias{gmdn}
\title{Calculate the Geothmetic Meandian}
\usage{
gmdn(..., tol = 1e-04)
}
\arguments{
\item{...}{A vector or sequence of numbers.}

\item{tol}{The tolerance for convergence.}
}
\value{
The geothmetic meandian.
}
\description{
Calculate the Geothmetic Meandian
}
